CA-305 Remove blocking code from process function and limit warning situation.
An infinite loop was added to force valid input on both A and B,
however there may be data left over from a previous iteration and
not future push will come. This code was removed completely, as it
was unnecessary.

A warning is printed out when the port supplying SRI does not
receive a new packet. This was changed to only print the warning
if no previous SRI was received from the port, because this may
happen quite often and there is not necessarily a problem.

Lastly, the left over cached data is cleared after being loaded
into the operational variables to avoid hanging around in certain
situations.